double

+++ src/main/java/cv/api/entity/OrderHisDetail.java
@@ -26,21 +26,21 @@ public class OrderHisDetail {
     @Column(name = "stock_code")
     private String stockCode;
     @Column(name = "qty", nullable = false)
-    private Float qty;
+    private Double qty;
     @Column(name = "unit")
     private String unitCode;
     @Column(name = "price", nullable = false)
-    private Float price;
+    private Double price;
     @Column(name = "amt", nullable = false)
-    private Float amount;
+    private Double amount;
     @Column(name = "loc_code")
     private String locCode;
     @Column(name = "weight")
-    private Float weight;
+    private Double weight;
     @Column(name = "weight_unit")
     private String weightUnit;
     @Column(name = "std_weight")
-    private Float stdWeight;
+    private Double stdWeight;
     @Transient
     private String userCode;
     @Transient
